adding spring forces.

+++ AeroliteEngine/src/Physics/Constants.h
@@ -5,6 +5,6 @@ constexpr int FPS = 60;
 constexpr int MILLISECS_PER_FRAME = 1000 / FPS;
 
 constexpr int PIXELS_PER_METER = 50;
-constexpr float GRAV_CONSTANT = 0.00000000000667430; // (N * m^2) / kg^2
+constexpr double GRAV_CONSTANT = 0.00000000000667430; // (N * m^2) / kg^2
 
 #endif
